+++ sqs.tf
@@ -4,33 +4,5 @@ module "inbound_queue" {
   organization = var.deployment_id
 }
 
-resource "aws_iam_policy" "inbound-sns-sqs-manage" {
-  tags   = local.default_tags
-  name   = "Etleap-inbound-sns-sqs-manage${local.resource_name_suffix}"
-  policy = <<EOF
-{
-  "Version": "2012-10-17",
-  "Id": "SqsAllowManagement",
-  "Statement": [
-  {
-    "Sid":"SqsAllowAll",
-    "Effect": "Allow",
-    "Action": "sqs:*",
-    "Resource": "${module.inbound_queue.sqs_queue_arn}"
-  },
-  {
-    "Sid":"SnsAllowAll",
-    "Effect": "Allow",
-    "Action": "sns:*",
-    "Resource": "${module.inbound_queue.sns_topic_arn}"
-  }
-]
-}
-EOF
-}
-
-resource "aws_iam_policy_attachment" "allow-sqs-app" {
-  name = "allow-sqs-app"
-  roles = [aws_iam_role.app.name]
-  policy_arn = aws_iam_policy.inbound-sns-sqs-manage.arn
-}
+# The permissions for the inbound queue are part of the app_various_limited
+# policy, to avoid breaching the 10 policy per role limit.